Rename sdk
