+++ packages/server/src/events/docUpdates/syncUsers.ts
@@ -1,7 +1,7 @@
 import { constants, logging } from "@budibase/backend-core"
 import { sdk as proSdk } from "@budibase/pro"
 import { DocUpdateEvent, UserGroupSyncEvents } from "@budibase/types"
-import { syncUsersToAllApps } from "../../sdk/app/applications/sync"
+import { syncUsersToAllApps } from "../../sdk/app/workspaces/sync"
 import { UpdateCallback } from "./processors"
 
 export default function process(updateCb?: UpdateCallback) {
