+++ packages/server/src/sdk/app/workspaces/index.ts
@@ -1,8 +1,8 @@
-import * as sync from "./sync"
-import * as utils from "./utils"
-import * as applications from "./applications"
 import * as imports from "./import"
 import * as metadata from "./metadata"
+import * as sync from "./sync"
+import * as utils from "./utils"
+import * as applications from "./workspaces"
 
 export default {
   ...sync,
